Restructured how the files are organized
Previously any folder that had source code went into src/
Now every script/plugin gets its own folder

I might put related scripts/plugins into the same folder in the future
when there's enough of them to warrant doing that. If you think the
files in this repo are too disorganized in the future and see this
commit then feel free to contact whoever is in control of this repo.